[heap, api] Treat instance embedder field conservatively with CppHeap
Previously, we would set embedder fields and do type checks (on
embedder fields) in the GC. This does not work nicely as embedder
fields contain system pointers whereas we can only operate with
tag-aligned reads/writes. The end result of assembling pointers was
somtimes broken for concurrent marking.

In this CL we treat any reference in the instance field as a
conservative reference for CppHeap. There's no checks required on
set/get at the expense of GC processing time.

Bug: chromium:1337690
Change-Id: I220efb76dec9f9329628a0d6e85647c1e2af92be