[compiler] fix incorrect string access in the background
Only Internalized and Thin strings should be accessed from the
background. My CL "[compiler] Remove map check in StringRef::length"
removed a check, and, while doing so, made it possible to read Cons
strings in the background, which is incorrect because such strings can
be mutated by the main threads.

Fixed: 1357438
Change-Id: I3a5445f91d10e5ceab62ac208a04be5ed71798ff