[csa][codegen] Optimize IsStrong/IsWeakOrCleared
The way to determine whether a MaybeObject is a strong or weak
reference to the heap object is to check its lowest two bits. However,
if the MaybeObject is known to not be a smi, that is, the lowest bit
is known to be 1, we can check one bit instead. This allows x64 to
select better instructions:

Before:
  movl r9,r11
  andl r9,0x3
  cmpb r9l,0x1

After:
  testb r11,0x2

Change-Id: I03623183406ad7d920c96a752651e0116a22832e