[heap] Update new space allocation counter during sweeping
We use an allocation observer to check when to start a minor GC. The
observer's step size is set as 80% of the new_space capacity.
Since scavenger allocates whenever it evacuates an object, the current
counter after a minor GC is the current new space size.
MinorMC needs to update the counter with the live bytes of sweeped pages
to maintain the same state.
Otherwise, we will trigger minor GCs to late.

Bug: v8:12612
Change-Id: I7de48665bbefa290ce336c3a31a2d724bb7fa53c