[heap] Decrease step size for ScavengeJob observer
Currently this observer uses 80% of new space capacity as step size.
But this means that after the first minor GC this will most likely
decouple from the current new space size.

Use 64K as step-size since this should be large enough to not cause
regression but it should still work for Scavenger and Minor MC such
that a step invocation will be performed close to reaching 80% of
new space capacity.

Bug: v8:12612
Change-Id: I4abc17eaeded90e0f72d9467a4410159ef0e6dda