A compositional approach on deciding heap limit
Currently, the approach to decide heap limit and when to gc consist of many different heuristic
that conflict each other (the 3% time rule, the caps on the memory, and idle time garbage collection),
requiring careful balancing.

Furthermore, we discovered that ultimately, no tuning of the heap limit rules are optimal,
as an optimal heap limit rules require that the extra memory for a heap
should be proportional to not the live set size,
but squareroot of the live set size.

Doing so will allocate memory sublinearly: compared to the original approach,
we take memory from the bigger heap, and give them to the smaller heap.

This correspond to a fundamental law of econmic: marginal diminishing return.

Taking memory from a big heap is not too noticable,
but giving the same amount to a small heap will cause a big change in heap limit:live set size.

This PR also disable the memory cap and the idle time garbage collection,
as their purpose is also achieved by the compositional heap limit rule.

Change-Id: Ie94fd2b487753b725f04398f07c106510f19a3e5