[maglev] Skip generator prologue for empty jump tables
We might generate a SwitchOnGeneratorState bytecode with zero jump table
entries if the JS code only has dead suspension points (where AST
suspensions are emitted, so suspend_count() > 0, but the bytecode for the
suspension ends up not being emitted because it's dead). An example
would be:

    async function() {
      return;
      await 0;
    }

In these cases, we can skip emitting the generator prologue, since the
function is not resumable.

Bug: v8:7700
Change-Id: Ie9f9d4fa8740f4ddc176cd5bbdc5beeda97ba8d5
Fixed: chromium:1370396