[heap]: Force promote unusable pages
A page is considered unusable if a GC occurs due to allocation failure
and there were no allocations on the page since the last GC. That
indicates that fragmentation on the page is such that remaining free
space could not be used. In such cases, we force promote the page.

Bug: v8:12612
Change-Id: I2d1fbb63bb4248559f23952f080235040cabe81a