[liveedit] Pass outer ScopeInfo for sloppy eval
The CL https://crrev.com/c/3932724 changed LiveEdit to replace the
ScopeInfo object of "unchanged" SFIs instead of updating the position
info in place. As a side-effect, we no longer set the outer ScopeInfo
on sloppy EVAL_SCOPEs.

This CL fixes that by properly passing the outer ScopeInfo along to
the compiler. It's also necessary to pass the outer ScopeInfo to the
reparse of the original script. Otherwise we'd get false positives in
the logic that detects outer scope changes for unchanged functions.

R=leszeks@chromium.org

Bug: chromium:1363561
Change-Id: Ia0bd8363cd835d7c7ffdc1d157c6657e6e2275c3