Remove builders which no longer exist.
These were removed in https://chromium-review.googlesource.com/c/chromium/src/+/3938463,
which triggered a latent bug in Buildbucket that lead to builds being
triggered with the wrong builder configuration.

Since these were experimental and had a comment next to them indicating
that they maybe provided little value, I've opted to remove them as a
mitigation, but it should be possible to re-add the new builders later,
if desired.

R=machenbach@google.com

Bug: 1372352, 1366151
Change-Id: I27ce23a984c93763de457d8b0fa26eece62f6e64