[wasm] Remove field to store "intermediate" error
We used to store the first error during function validation there, and
then later return it instead of the outer (module-wide) decoding error.
This is redundant, we can just directly store a function validation
error in the outer error field.
This also removes the boolean parameter to {Decoder::FinishDecoding}.
This was always redundant, since the "intermediate error" was only set
if function validation was enabled before.

R=jkummerow@chromium.org

Bug: v8:13371
Change-Id: Ib812d4ba0ea10e9a5dda46cac4e33d427dd9c16c