Remove remaining reclient shadow builders (goma cache silo + Mac64)
The Mac64 shadow is safe to remove as the multibranch builder default was already set to reclient https://crrev.com/c/3935923.

For consistency, in this CL we also set GOMA.NO and RECLIENT.DEFAULT explicitly on the existing Mac64 builder. We can observe that this does not affect the luci generated config.

This can then be safely cleaned up in a followup CL.

Last migration: non goma cache silo https://crrev.com/c/3935511

Bug: b:238274944
Change-Id: Ia1b6262030ebefdaad33e19278f29b7be77bfb64