[base] Implement SharedMutex via std::shared_mutex
On Mac we already use {std::shared_mutex} to implement
{base::SharedMutex}, as {pthread_rwlock_t} causes deadlocks if combined
with signals. Other platforms still use such low-level primitives.

In order to streamline the implementation, and as a preparation to use
{std::shared_mutex} directly everywhere, we make {base::SharedMutex} a
wrapper around {std::shared_mutex} on all platforms.

If this causes regressions anywhere, it will be reverted and the switch
to {std::shared_mutex} will be abandoned.

R=ishell@chromium.org

Bug: v8:13256
Change-Id: I39c41666451884b974f8e5e592df0a36e8e80445