[test] Fix optimization marking for in-progress
%OptimizeFunctionOnNextCall could be called while compilation (Maglev or
Turbofan) was still on-going -- this would make DCHECKs fail when
attempting to save the compiled code on the feedback vector. This is not
an issue normally when tests are carefully crafted, but can easily
happen during fuzzing.

Unify the two %Optimize*OnNextCall paths, and make sure that they check
for in-progress compilation.

Fixed: v8:13404
Change-Id: I021e1305bc2ba769b7484f9a122792aeb22060ac