[interpreter] Reuse predefined temp register when build ForOf statement
When build the for...of statement with declarations, parser will create
a temporary .for register to store the each variable. During bytecode
generation, the current implementation will allocate a new temporary
register to pass the value which is redundant. This patch will try
reuse the predefined temp register to reduce a move when possible.

Change-Id: Ia44c6e81dd7ec1dbd22a8861e52642aa5a556ce3