[runtime] implement redesign of [[GetOwnPropertyDescriptor]]
This new implementation includes a CSA handle for the Proxy
[[GetOwnPropertyDescriptor]] case, and adds a new code stub for
dispatching to the appropriate handler.

This results in a significant performance increase on a customer
application which makes significant use of GetOwnPropertyDescriptor on
Proxy objects during the application startup.

We have found minimal impact on industry standard benchmarks such as
Speedometer 2.0.

Change-Id: I45f8a9190917938b635bfa58145902f4a0da12fc