[maglev] Extract out DeoptFrame from DeoptInfo
Extract out a concept of a DeoptFrame from DeoptInfo, which separates
the frame state (like bytecode offset and registers) from deopt
information (like reason and PC).

The DeoptFrame is additionally subclassed to a separate
InterpretedDeoptFrame (with some tagged union magic rather than 'proper'
subclassing so that in the future all DeoptFrames are the same size and
aren't truncated by casting). This way we can add different frames in
the future, in particular builtin continuation frames.

Also this cleans up parent walks, since we no longer walk the caller
state and compilation unit separately.

Bug: v8:7700
Change-Id: I1cecb3ae805c55235b6d74ec114d72de98d3751e