Rename compilator trybots
This drops the _ng infix from compilator names, which existed only
to ease the roll-out and to not clash with existing trybots.

By now all existing trybots were merged into compilators or were
renamed. This is also prepared by setting the compilator name as
a property and by preparing the recipes to accept both types of
names (old and new) for source side look-ups.

Bug: chromium:890222
Change-Id: I14a1a48d5449f6d7a15ee12b4aadebafbe0ee510