[wasm][decoder] Skip the name parameter if it is unused
In release builds the name will only be used for producing an error
message. For decoder instantiations that do not run full validation, it
is thus never used. Eliminating the unused parameter (by passing an
empty object instead) saves >3kB of binary size on x64, and also makes
the decoder slightly faster.

Drive-by: Inline {validate_size} into the only caller to simplify the
code.

R=ahaas@chromium.org

Change-Id: If50a118789c3e9c56ee24b68f35fdd3920e59e9e